# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtual PC VM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VmArchitecture =
  (typeof VmArchitecture)[keyof typeof VmArchitecture];

export const VmArchitecture = {
  arm: "arm",
  arm64: "arm64",
  x86: "x86",
  x86_64: "x86_64",
  i386: "i386",
  powerpc: "powerpc",
} as const;

export type VmMachineType = (typeof VmMachineType)[keyof typeof VmMachineType];

export const VmMachineType = {
  q35: "q35",
  pc: "pc",
  virt: "virt",
  isapc: "isapc",
} as const;

export type VmFirmware = (typeof VmFirmware)[keyof typeof VmFirmware];

export const VmFirmware = {
  bios: "bios",
  uefi: "uefi",
} as const;

export type VmStatus = (typeof VmStatus)[keyof typeof VmStatus];

export const VmStatus = {
  stopped: "stopped",
  running: "running",
  starting: "starting",
  error: "error",
} as const;

export interface Vm {
  id: number;
  name: string;
  ramMb: number;
  cpuCores: number;
  architecture: VmArchitecture;
  machineType: VmMachineType;
  firmware: VmFirmware;
  enableMtcg: boolean;
  enableKvm: boolean;
  disableTsc: boolean;
  isoPath?: string | null;
  diskId?: number | null;
  status: VmStatus;
  createdAt: string;
}

export type CreateVMRequestArchitecture =
  (typeof CreateVMRequestArchitecture)[keyof typeof CreateVMRequestArchitecture];

export const CreateVMRequestArchitecture = {
  arm: "arm",
  arm64: "arm64",
  x86: "x86",
  x86_64: "x86_64",
  i386: "i386",
  powerpc: "powerpc",
} as const;

export type CreateVMRequestMachineType =
  (typeof CreateVMRequestMachineType)[keyof typeof CreateVMRequestMachineType];

export const CreateVMRequestMachineType = {
  q35: "q35",
  pc: "pc",
  virt: "virt",
  isapc: "isapc",
} as const;

export type CreateVMRequestFirmware =
  (typeof CreateVMRequestFirmware)[keyof typeof CreateVMRequestFirmware];

export const CreateVMRequestFirmware = {
  bios: "bios",
  uefi: "uefi",
} as const;

export interface CreateVMRequest {
  name: string;
  ramMb: number;
  cpuCores: number;
  architecture: CreateVMRequestArchitecture;
  machineType: CreateVMRequestMachineType;
  firmware: CreateVMRequestFirmware;
  enableMtcg: boolean;
  enableKvm: boolean;
  disableTsc: boolean;
  isoPath?: string | null;
  diskId?: number | null;
}

export type VMStatusResponseStatus =
  (typeof VMStatusResponseStatus)[keyof typeof VMStatusResponseStatus];

export const VMStatusResponseStatus = {
  stopped: "stopped",
  running: "running",
  starting: "starting",
  error: "error",
} as const;

export interface VMStatusResponse {
  id: number;
  status: VMStatusResponseStatus;
  message: string;
}

export type VirtualDiskFormat =
  (typeof VirtualDiskFormat)[keyof typeof VirtualDiskFormat];

export const VirtualDiskFormat = {
  qcow2: "qcow2",
  raw: "raw",
  vmdk: "vmdk",
  vdi: "vdi",
} as const;

export interface VirtualDisk {
  id: number;
  name: string;
  sizeMb: number;
  format: VirtualDiskFormat;
  path: string;
  createdAt: string;
}

export type CreateDiskRequestFormat =
  (typeof CreateDiskRequestFormat)[keyof typeof CreateDiskRequestFormat];

export const CreateDiskRequestFormat = {
  qcow2: "qcow2",
  raw: "raw",
  vmdk: "vmdk",
  vdi: "vdi",
} as const;

export interface CreateDiskRequest {
  name: string;
  sizeMb: number;
  format: CreateDiskRequestFormat;
}