# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Virtual PC VM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /disks:
    get:
      operationId: listDisks
      tags: [disks]
      summary: List all virtual disks
      responses:
        "200":
          description: List of disks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualDisk"
    post:
      operationId: createDisk
      tags: [disks]
      summary: Create a new virtual disk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiskRequest"
      responses:
        "201":
          description: Disk created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualDisk"
  /disks/{id}:
    delete:
      operationId: deleteDisk
      tags: [disks]
      summary: Delete a virtual disk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disk deleted
components:
  schemas:
...
    VirtualDisk:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sizeMb:
          type: integer
        format:
          type: string
          enum: [qcow2, raw, vmdk, vdi]
        path:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - sizeMb
        - format
        - path
        - createdAt
    CreateDiskRequest:
      type: object
      properties:
        name:
          type: string
        sizeMb:
          type: integer
        format:
          type: string
          enum: [qcow2, raw, vmdk, vdi]
      required:
        - name
        - sizeMb
        - format