# lib/api-zod/src/generated/types/createDiskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtual PC VM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDiskRequestFormat } from "./createDiskRequestFormat";

export interface CreateDiskRequest {
  name: string;
  sizeMb: number;
  format: CreateDiskRequestFormat;
}